feat: derive Get Directions link in locator from ref_listings if present (#860)
Updates the locator to derive the Get Directions link on result cards
from the ref_listings field in the search response if present, otherwise
fallback to the yextDisplayCoordinate. If that is also not present,
don't display the get direction button.

Copies the logic in #846 to determine the URL.

J=WAT-5136
TEST=manual

Created a locator in the Krusty dev account (514581), which has some
entities with ref_listings data pointing to simulated URLs. Confirmed
that when the ref_listings field in the search response had a listingUrl
for a GOOGLEMYBUSINESS publisher, the Get Directions button pointed to
the listingUrl rather than the Google Maps one generated from the
yextDisplayCoordinate field. Confirmed that when the ref_listings field
was empty or did not have a GOOGLEMYBUSINESS publisher, the
yextDisplayCoordinate field was used to generate the Get Direction URL
instead.

jam:https://jam.dev/c/a41f5a6d-1491-4a7a-96aa-936766e204e6