fix: locator migrations (#865)
The puck name of the locator is `Locator`, which is determined by the
`LocatorCategoryComponents` object in
`packages/visual-editor/src/components/categories/LocatorCategory.tsx`,
which is copied into the puck config.

When I added the Locator tests a few months ago, I missed that we
renamed `LocatorComponent` to `Locator` in the config.

So, the component tests were expecting migrations to operate on
`LocatorComponent` but the real pages were expecting `Locator`. We
previously had a small locator migration, but it just set a non-breaking
default value so the mistake didn't show up.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>