chore: remove types from migrations (#875)
The migration props need to be frozen with the current prop shape. If
they remain typed, future updates to the prop shape will fail the build.

If we need to modify the migrations as part of bug bash fixes, we can
temporarily add the types back for the affected migration