fix: slots performance when tab is switched (#883)
Fixes a variety of inefficiencies that became more noticeable because of
the extra overhead of slots

1. The useEntityFields and useDocument contexts were always updating
when receiving message, causing every component to re-render. This has
been fixed by checking if the entity fields or document actually updated
2. The LayoutSaveState was also always updating when receiving a message
even if the history hadn't changed
4. We need to do `resolveAllData` when the document changes but this was
occurring in the LayoutHeader, which was remounting frequently due to
other updates. The resolve has been moved to its own override.
5. The image object reference was not stable, causing the images to
re-render more than necessary

The editor can still freeze up for a couple seconds if you leave the tab
while a save/publish is in progress but otherwise does well. I've added
our context to a Puck issue about slot performance.

Test site:
https://www.yext.com/s/4259018/yextsites/159383/branches/131432/editor#pageSetId=locations&locale=en&themeId=pages-visual-editor-starter-other&entityId=1095527849