feat: adding comma seperated prop for TextList Component (#902)
On professional VLE pages (location templates), we’re following the new
two-column layout, where, the left side shows the description, while the
right side lists details like products, services, and specialties. Right
now, the right column looks too large because the list is rendered as a
`<ul>`. To fix this, there will be a new prop `Comma Separated` on the
TextList component. When toggled on, the list will display as a
comma-separated string instead of a bulleted list. By default, this prop
stays off.



https://github.com/user-attachments/assets/d47962f9-3f97-4717-ae3e-4fdc75207da5

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>