fix: handle insight section react issues and RTF's logic (#906)
For the react issues, there was this
[error](https://react.dev/errors/185?invariant=185) which doesn't occur
anymore.

For the RTF logic change, even if we deleted our RTF's value, the RTF
data looks like html: '' and json: '', so the value appears to be a rtf
element and it returns `<div w/ rtf theme classes></div>`. This fixes a
bunch of places where if you delete the text on BodyTextSlot, the slot
isn't clickable anymore (in the editor).

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>