fix: update default phone numbers (#913)
Our phone number input component provides us with unformatted phone
numbers with a leading `+` and country code, which is the required
format for the phone number parser we use. Our default phone numbers,
however, were pre-formatted as US domestic numbers and failed to get
successfully formatted before displaying, making the
Domestic/International toggle not work until the number was edited.

This removes the unnecessary formatting, adds the `+1` country code, and
also cleans the phone number before attempting to parse in order to
prevent this from occurring again in the future.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>