fix: don't allow user settable link prop for getDirections CTA (#920)
Tested in platform, in the Hero component set a text and link value,
then switched to Get Directions, updates the page and verified it used
the coordinate field. Then, switched back to text and link and verified
that the prop value was still set and functioned correctly. Also
confirmed that Address component still functioned as expected with the
address link.

Confirmed that these are the only places that need updating, Address and
Core Info still correctly pass in the addressLink as the link prop.

Test site:
https://www.yext.com/s/4259018/yextsites/161656/branches/131864/editor#pageSetId=cta-test&locale=en&themeId=ctatest&entityId=1095527849

Co-authored-by: coltondemetriou <coltondemetriou@gmail.com>