fix: logic in PromoSection slots migration (#931)
Some of the logic was a bit off as we were removing the ability to have
constantValueOverrides in PromoSection as well as slotifying. Adjusted
and confirmed with the existing data in actual accounts that the
migration now works, added the previously not properly migrated data as
a test case as well.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>