fix: shared styles lost when new card is added (#938)
When a new card was added to a card section (either via adding in the
constant value or via a new list item on the entity), the shared card
styles were being reset to the defaults. This was because the new card
was being created with default props, triggering it to update the shared
context, triggering the existing cards to dispatch updates. This change
makes it so that new cards are initialized with the shared styling props
of the already existing cards.