fix: tab slowness (#940)
#935 added a deepEqual check for re-resolving all data which seemed to
have helped a bit.

This PR adds additional save state management, which helps particularly
when clearing local changes.
Reducing the placeholder image size seems to make the re-rendering a
little faster.

There still is a delay when publishing and then switching tabs. I'm not
entirely sure what is triggering it but the slowness seems to be on
Puck's side. Puck has some slot performance improvements coming up, so I
think we should take a look again after those are done.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>