fix: fixed header rendering flash and product card image ratio (#946)
- Product card images appear differently in editor than live page
  - Fix: different style applied to the Slot
- Secondary header links appear left-aligned in editor but right-aligned
on live page
  - Fix: different style applied to the Slot 
- Directory heading alignment not being respected on live page
   - Fix: different style applied to the Slot
- Fixed position header content flash on live page
- Cause: We had a placeholder div to move the top of the page down below
the header. However, this required JS to calculate the height of the
header and so it would load as 0 and then update to be correct
- Fix: css position 'sticky' is supposed to handle this. It previously
did not work because the height for the whole page was set to 100vh (in
order to push the footer to the bottom of the page). By changing this to
min-height, the footer remains always at the bottom but 'sticky' for the
header works as expected for the whole length of the page. So, I think
we can consolidate 'fixed' into 'sticky'
- Example with the new sticky:
https://daintily-tops-rooster.pgsdemo.com/dc/washington/2517-k-street-northwest

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>